sizes are Int instead of Natural

+++ src/HLinear/Hook/EchelonTransformation/Column.hs
@@ -65,7 +65,7 @@ toVector
   :: Ring a
   => EchelonTransformationColumn a -> Vector a
 toVector (EchelonTransformationColumn o v) =
-  v `V.snoc` MS.one <> V.replicate (fromIntegral o) zero
+  v `V.snoc` MS.one <> V.replicate o zero
 
 --------------------------------------------------------------------------------
 -- Eq, Show, and NFData instancs
